Move Estimator to estimators package  * Move SubGraph to libs  * Move Summarizer to models  * Update docs  * Update examples
