+++ tests/experiments/test_subgraph.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, division, print_function
 
 import tensorflow as tf
 import polyaxon as plx
-from polyaxon.experiments import SubGraph
+from polyaxon.libs.subgraph import SubGraph
 from polyaxon.libs.configs import SubGraphConfig
